fix(filehandling): create new file before opening it
To avoid/workaround #7 the new file is created directly and then opened.
This avoids the confusion with being asked for the new file name and
on the other side having an "untitled:.." tab.
Added an error message to the user as well.
Added the missing type for new fba.

Issue: #7