feat(restquery): re-trigger badge queries on dlt doc view changes
If the active dlt document changes the badges will be re-queried.

This requires DLT-logs v1.3.0 to work.
There is one problem remaining that on new DLT documents the restQuery
is triggered so early that the document isn't parsed yet.
Could delay the queries here but will investigate for a change in DLT-logs extension.