feat(badges): show as well for errors
Prev. the badges have only been shown on unchecked root causes.
Changed that to hide the badges only if it's marked as 'ok'.
So especially keep on showing it if its marked as 'error' as the badges
are frequently used to provide additional infos (e.g. versions, type of problem,
...).