fix: support save as / copy file use-case
The old implementation assumed that the fbUids are
unique above all items.
But due to e.g. copy file that's not the case.
They are only unique (supposed to) within a single
fba document.
Used the authority for fbaFs provider to contain
the orig document uri (hex encoded to fit as authority).
Prepared as well to change for new v5 of short-unique-id.