feat(fai): prompt files and log to output channel
Added support for prompt files.
They are dynamically searched in the same dir as the fishbones opened
and need to have the name: "<usercmd>.prompt.md.

E.g. if you name it "analyse.prompt.md" it will provide the prompt for the analyse
command.

The file format needs to be the same as
https://code.visualstudio.com/docs/copilot/copilot-customization#_prompt-file-structure

So starting with a front matter header
---
mode: 'ask'
model: <model to use>
tools: [list of tools to use]
description: ...
---
<prompt as markdown>.

Currently all data from the header is ignored but support will be
added later.
The intention is that tools, model... overwrite the default settings.
So if not provided defaults will be used.

Logs are now directed to the output channel named "Fishbone".
With log level "debug" all communication with the LLM is shown.